Fixed bugs with updating summary records and some spelling errors.
In particular, when deciding whether to update a summary record with
more recent data, I was matching on summary_id rather than the start time.
This required updating the logic in the method used to compare API responses
to the database data, as well as changing some arguments from the defaults
in API endpoints that don't use the standard way of recording time.