Correct database update logic for incoming Wellness API responses.
The original update_db_from_api_response had operated under the assumption
that 'summaryId' was a suitable primary key, and that ordering based on duration
was suitable. It turned out these assumptions were incorrect; multiple summaryId's
can exist for the same activity, and, in some cases, the same summaryId and duration
can be sent in a response with different data.

The function was updated to pull the primary key and class object directly from the
object that was passed in (eliminating the `match_attr` and `table_obj` parameters).
`order_attr` is now a required parameter.

The API endpoint handlers and test suite was updated to reflect these changes as well.