refactor configuration handling, moving db intialization and session_scoping to rely entirely on app's configuration.
Prior to this commit, the sqlalchemy `Engine` was declare in the db module.

This meant there was no possibility for dynamically changing what engine session_scope was bound to, and made testing finicky. The goal of this change is to allow each test to initialize its own database and allow the tests to run in parallel.